Fixed original rain effect to start immedately instead of waiting up to three seconds.  Also changed some fields to float to avoid the popping of new drops at the top.

+++ src/OriginalRainEffect.h
@@ -14,11 +14,11 @@ private:
 
 	struct RainDrop
 	{
-		uint32_t birth;
+		float birth;
 		uint8_t strip;
 		float hue;
 		float saturation;
-		uint8_t tail;
+		float tail;
 	};
 
 	struct Hsv {
